+++ modules/database/tools.py
@@ -22,3 +22,11 @@ def extract_sql_query(model_response: str) -> str:
         return sql_query
 
     return "No valid SQL query found in the response."
+
+
+def convert_list_to_string(results, columns):
+    """Convert list of tuples to string with column names"""
+    results_str = f"COLUMNS:{','.join(columns)}\n"
+    for row in results:
+        results_str += str(row) + "\n"
+    return results_str
